chore: arrange folders and fallback to bot if no command
